refactor: create pkg for wait logic
